+++ app/src/main/java/com/teamtriad/forpets/ui/transport/ViewPagerVolListFragment.kt
@@ -15,11 +15,12 @@ import com.teamtriad.forpets.R
 import com.teamtriad.forpets.databinding.FragmentViewPagerVolListBinding
 import com.teamtriad.forpets.model.tmp.Volunteers
 import com.teamtriad.forpets.ui.transport.adpater.VolListRecyclerViewAdapter
-import com.teamtriad.forpets.ui.transport.bottomSheetDialog.LocationPickerForFragmentManager
 import com.teamtriad.forpets.util.formatDate
 import com.teamtriad.forpets.util.formatDateWithYear
 import java.util.Calendar
 import java.util.TimeZone
+
+private const val LOCATION = "onlyCounty"
 
 class ViewPagerVolListFragment : Fragment() {
 
@@ -70,11 +71,17 @@ class ViewPagerVolListFragment : Fragment() {
             }
 
             tietFrom.setOnClickListener {
-                showModalBottomSheet()
+                val action = TransportListsFragmentDirections
+                    .actionTransportListsFragmentToLocationPickerForNavigation(LOCATION)
+
+                findNavController().navigate(action)
             }
 
             tietTo.setOnClickListener {
-                showModalBottomSheet()
+                val action = TransportListsFragmentDirections
+                    .actionTransportListsFragmentToLocationPickerForNavigation(LOCATION)
+
+                findNavController().navigate(action)
             }
 
             efabVolList.setOnClickListener {
@@ -82,14 +89,6 @@ class ViewPagerVolListFragment : Fragment() {
                     .navigate(R.id.action_transportListsFragment_to_transportVolFragment)
             }
         }
-    }
-
-    private fun showModalBottomSheet() {
-        val bottomSheet = LocationPickerForFragmentManager()
-        bottomSheet.show(
-            requireActivity().supportFragmentManager,
-            LocationPickerForFragmentManager.TAG
-        )
     }
 
     private fun setDatePicker() {
